machine-learning-models-main/requirements.txt
streamlit
tensorflow
numpy
matplotlib
pillow